Extract spectrum.polling
